+++ src/main/java/com/ezen/management/service/LessonService.java
@@ -11,6 +11,6 @@ public interface LessonService {
 
     List<Lesson> findAllGreaterThan(LocalDate day);
 
-    Lesson findById(int lessonIdx);
+    Lesson findById(Long lessonIdx);
 
 }
